dapla-manual/blog/posts/2023-05-02-store-datasett-dapla/benchmark_code/pyproject.toml
[tool.poetry]
name = "benchmark-code"
version = "0.1.0"
description = ""
authors = ["Michael Moen Allport <mmallport@gmail.com>"]
readme = "README.md"
packages = [{ include = "benchmark_code", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4"
dapla-toolbelt = ">=2.0.13"
pandas = ">=2.2.2"
polars = ">=0.20.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
